+++ src/main/java/Teacher/TeacherAttentionController.java
@@ -2,20 +2,17 @@ package Teacher;
 
 import Application.Attention;
 import Application.AttentionDAO;
+import Utils.DialogUtils;
 import com.jfoenix.controls.JFXDatePicker;
 import com.jfoenix.controls.JFXTimePicker;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.DatePicker;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 
 import java.sql.SQLException;
 import java.util.Date;
+import java.util.Optional;
 
 public class TeacherAttentionController {
 
@@ -54,42 +51,52 @@ public class TeacherAttentionController {
     private TextArea teacher_attention_text_area;
 
     @FXML
-    void teacherAttentionAdd(ActionEvent event){
+    void teacherAttentionAdd(ActionEvent event) {
 
-            AttentionDAO.insertAttention(teacher_attention_text_area.getText(),teacher_attention_time_picker.getValue().toString() + " " +teacher_attention_date_picker.getValue().toString());
+        if (DialogUtils.addDialog().get() == ButtonType.OK) {
+
+            AttentionDAO.insertAttention(teacher_attention_text_area.getText(), teacher_attention_time_picker.getValue().toString() + " " + teacher_attention_date_picker.getValue().toString());
             searchAttentions();
+        }
 
     }
+
     @FXML
-    void searchAttentions(){
+    void searchAttentions() {
         ObservableList<Attention> attentions = AttentionDAO.searchAttentions();
         teacher_attention_table_view.setItems(attentions);
     }
 
 
     @FXML
-    void teacherAttentionDelete(ActionEvent event){
-        int x = Integer.parseInt(teacher_attention_id_text_field.getText());
-        AttentionDAO.deleteAttentionWithId(x);
-        searchAttentions();
+    void teacherAttentionDelete(ActionEvent event) {
+
+        if (DialogUtils.deleteDialog().get() == ButtonType.OK) {
+            int x = Integer.parseInt(teacher_attention_id_text_field.getText());
+            AttentionDAO.deleteAttentionWithId(x);
+            searchAttentions();
+        }
     }
 
     @FXML
-    void teacherAttentionUpdate(ActionEvent event){
-        int x = Integer.parseInt(teacher_attention_id_text_field.getText());
-        AttentionDAO.updateAttention(x, teacher_attention_text_area.getText(),teacher_attention_time_picker.getValue().toString() + " " + teacher_attention_date_picker.getValue().toString());
-        searchAttentions();
+    void teacherAttentionUpdate(ActionEvent event) {
+
+        if (DialogUtils.updateDialog().get() == ButtonType.OK) {
+            int x = Integer.parseInt(teacher_attention_id_text_field.getText());
+            AttentionDAO.updateAttention(x, teacher_attention_text_area.getText(), teacher_attention_time_picker.getValue().toString() + " " + teacher_attention_date_picker.getValue().toString());
+            searchAttentions();
+        }
 
     }
 
     @FXML
-    public void initialize(){
+    public void initialize() {
 
         searchAttentions();
 
-        teacher_attention_id_column.setCellValueFactory(cellData->cellData.getValue().student_idProperty().asObject());
-        teacher_attention_column.setCellValueFactory(cellData->cellData.getValue().attentionProperty());
-        teacher_attention_date_column.setCellValueFactory(cellData->cellData.getValue().dateProperty());
+        teacher_attention_id_column.setCellValueFactory(cellData -> cellData.getValue().student_idProperty().asObject());
+        teacher_attention_column.setCellValueFactory(cellData -> cellData.getValue().attentionProperty());
+        teacher_attention_date_column.setCellValueFactory(cellData -> cellData.getValue().dateProperty());
     }
